Lint this, lint that
